# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: متجري — SaaS perfume & cosmetics storefront platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get merchant dashboard stats — orders this month, revenue, top products
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Dashboard Settings ───────────────────────────────────
  /dashboard/settings:
    get:
      operationId: getDashboardSettings
      tags: [dashboard]
      summary: Get merchant store settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateDashboardSettings
      tags: [dashboard]
      summary: Update merchant store settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Products ─────────────────────────────────────────────
  /dashboard/products:
    get:
      operationId: listProducts
      tags: [dashboard]
      summary: List merchant products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [dashboard]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/products/{id}:
    get:
      operationId: getProduct
      tags: [dashboard]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [dashboard]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [dashboard]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Orders ───────────────────────────────────────────────
  /dashboard/orders:
    get:
      operationId: listOrders
      tags: [dashboard]
      summary: List merchant orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /dashboard/orders/{id}:
    get:
      operationId: getOrder
      tags: [dashboard]
      summary: Get a single order with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [dashboard]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Discount Codes ───────────────────────────────────────
  /dashboard/discounts:
    get:
      operationId: listDiscounts
      tags: [dashboard]
      summary: List merchant discount codes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscountCode"
    post:
      operationId: createDiscount
      tags: [dashboard]
      summary: Create a discount code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscountCode"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/discounts/{id}:
    delete:
      operationId: deleteDiscount
      tags: [dashboard]
      summary: Delete a discount code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /dashboard/discounts/{id}/toggle:
    patch:
      operationId: toggleDiscount
      tags: [dashboard]
      summary: Toggle discount code active/inactive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscountCode"

  # ─── Storefront (Public) ──────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── Auth ───────────────────────────────────────────────
...
    Merchant:
      type: object
      required: [id, slug, storeName, email, accentColor, createdAt]
      properties:
        id:
          type: number
        slug:
          type: string
        storeName:
          type: string
        email:
          type: string
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        accentColor:
          type: string
        bankTransferInfo:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
        productCount:
          type: ["number", "null"]
        orderCount:
          type: ["number", "null"]
    MerchantUpdate:
      type: object
      properties:
        storeName:
          type: string
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        accentColor:
          type: string
        bankTransferInfo:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]

    # ─── Dashboard ──────────────────────────────────────────
    DashboardStats:
      type: object
      required: [ordersThisMonth, revenueThisMonth, totalOrders, totalRevenue, topProducts, ordersByStatus, recentOrders]
      properties:
        ordersThisMonth:
          type: number
        revenueThisMonth:
          type: number
        totalOrders:
          type: number
        totalRevenue:
          type: number
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/TopProduct"
        ordersByStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
    TopProduct:
      type: object
      required: [productId, productName, totalSold, totalRevenue]
      properties:
        productId:
          type: number
        productName:
          type: string
        totalSold:
          type: number
        totalRevenue:
          type: number
    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: number

    # ─── Products ───────────────────────────────────────────
    ProductVariant:
      type: object
      required: [id, variantLabel, price, stock]
      properties:
        id:
          type: number
        variantLabel:
          type: string
        price:
          type: number
        stock:
          type: number
    ProductVariantInput:
      type: object
      required: [variantLabel, price]
      properties:
        variantLabel:
          type: string
        price:
          type: number
        stock:
          type: number
    Product:
      type: object
      required: [id, merchantId, name, category, isActive, createdAt, variants]
      properties:
        id:
          type: number
        merchantId:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        noteTop:
          type: ["string", "null"]
        noteHeart:
          type: ["string", "null"]
        noteBase:
          type: ["string", "null"]
        skinType:
          type: ["string", "null"]
        ingredients:
          type: ["string", "null"]
        batchExpiry:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
    ProductInput:
      type: object
      required: [name, category, variants]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        noteTop:
          type: ["string", "null"]
        noteHeart:
          type: ["string", "null"]
        noteBase:
          type: ["string", "null"]
        skinType:
          type: ["string", "null"]
        ingredients:
          type: ["string", "null"]
        batchExpiry:
          type: ["string", "null"]
        isActive:
          type: boolean
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariantInput"
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        noteTop:
          type: ["string", "null"]
        noteHeart:
          type: ["string", "null"]
        noteBase:
          type: ["string", "null"]
        skinType:
          type: ["string", "null"]
        ingredients:
          type: ["string", "null"]
        batchExpiry:
          type: ["string", "null"]
        isActive:
          type: boolean
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariantInput"

    # ─── Orders ─────────────────────────────────────────────
    OrderItem:
      type: object
      required: [id, productName, variantLabel, quantity, priceAtOrder]
      properties:
        id:
          type: number
        variantId:
          type: ["number", "null"]
        productName:
          type: string
        variantLabel:
          type: string
        quantity:
          type: number
        priceAtOrder:
          type: number
    Order:
      type: object
      required: [id, merchantId, customerName, customerPhone, customerAddress, paymentMethod, isGift, subtotal, total, status, createdAt]
      properties:
        id:
          type: number
        merchantId:
          type: number
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        paymentMethod:
          type: string
        isGift:
          type: boolean
        giftMessage:
          type: ["string", "null"]
        discountCode:
          type: ["string", "null"]
        subtotal:
          type: number
        total:
          type: number
        status:
          type: string
        createdAt:
          type: string
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
    OrderListResponse:
      type: object
      required: [orders, total, page, pageSize]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: number
        page:
          type: number
        pageSize:
          type: number
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
...
    DiscountCode:
      type: object
      required: [id, merchantId, code, percentOff, isActive, createdAt]
      properties:
        id:
          type: number
        merchantId:
          type: number
        code:
          type: string
        percentOff:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
    DiscountInput:
      type: object
      required: [code, percentOff]
      properties:
        code:
          type: string
        percentOff:
          type: number
        isActive:
          type: boolean
...