# lib/api-zod/src/generated/types/discountCode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * متجري — SaaS perfume & cosmetics storefront platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DiscountCode {
  id: number;
  merchantId: number;
  code: string;
  percentOff: number;
  isActive: boolean;
  createdAt: string;
}