# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * متجري — SaaS perfume & cosmetics storefront platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  merchantId: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  paymentMethod: string;
  isGift: boolean;
  /** @nullable */
  giftMessage?: string | null;
  /** @nullable */
  discountCode?: string | null;
  subtotal: number;
  total: number;
  status: string;
  createdAt: string;
}