# lib/api-zod/src/generated/types/orderDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * متجري — SaaS perfume & cosmetics storefront platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';
import type { OrderItem } from './orderItem';

export type OrderDetail = Order & {
  items: OrderItem[];
};