# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * متجري — SaaS perfume & cosmetics storefront platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: number;
  /** @nullable */
  variantId?: number | null;
  productName: string;
  variantLabel: string;
  quantity: number;
  priceAtOrder: number;
}