# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * متجري — SaaS perfume & cosmetics storefront platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductVariantInput } from './productVariantInput';

export interface ProductInput {
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  imageUrls?: string[];
  /** @nullable */
  noteTop?: string | null;
  /** @nullable */
  noteHeart?: string | null;
  /** @nullable */
  noteBase?: string | null;
  /** @nullable */
  skinType?: string | null;
  /** @nullable */
  ingredients?: string | null;
  /** @nullable */
  batchExpiry?: string | null;
  isActive?: boolean;
  variants: ProductVariantInput[];
}